requirements.txt
discord.py==2.4.0
flask==3.1.0
pyyaml==6.0.1